# docs/callback/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AI Agent Productivity Platform - Callback",
        "contact": {},
        "version": "v0.0.1"
    },
    "basePath": "/callback/v1",
    "paths": {
        "/api/category/info": {
            "get": {
                "description": "获取Maas平台知识库信息（模型扩展调用）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "获取Maas平台知识库信息（模型扩展调用）",
                "parameters": [
                    {
                        "description": "根据知识库名称请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SearchKnowledgeInfoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/docstatus": {
            "post": {
                "description": "更新文档状态（模型扩展调用）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "更新文档状态（模型扩展调用）",
                "parameters": [
                    {
                        "description": "更新文档状态请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CallbackUpdateDocStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/model/{modelId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "根据ModelId获取模型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型ID",
                        "name": "modelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ModelInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/model/{modelId}/chat/completions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "Model Chat Completions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型ID",
                        "name": "modelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mp_common.LLMReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mp_common.LLMResp"
                        }
                    }
                }
            }
        },
        "/model/{modelId}/embeddings": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "Model Embeddings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型ID",
                        "name": "modelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mp_common.EmbeddingReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mp_common.EmbeddingResp"
                        }
                    }
                }
            }
        },
        "/model/{modelId}/gui": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "Model Gui",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型ID",
                        "name": "modelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mp_common.GuiReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mp_common.GuiResp"
                        }
                    }
                }
            }
        },
        "/model/{modelId}/ocr": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "Model Ocr",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型ID",
                        "name": "modelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mp_common.OcrResp"
                        }
                    }
                }
            }
        },
        "/model/{modelId}/pdf-parser": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "Model PdfParser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型ID",
                        "name": "modelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件名",
                        "name": "file_name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mp_common.PdfParserResp"
                        }
                    }
                }
            }
        },
        "/model/{modelId}/rerank": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "Model Rerank",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模型ID",
                        "name": "modelId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mp_common.RerankReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mp_common.RerankResp"
                        }
                    }
                }
            }
        },
        "/workflow/list": {
            "get": {
                "description": "根据userId和spaceId获取Workflow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "根据userId和spaceId获取Workflow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "获取工作流参数userId",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "获取工作流参数orgId",
                        "name": "orgId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/workflow/tool/custom": {
            "get": {
                "description": "获取自定义工具详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "获取自定义工具详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "customToolId",
                        "name": "customToolId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.CustomToolDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/workflow/tool/square": {
            "get": {
                "description": "获取内置工具详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "callback"
                ],
                "summary": "获取内置工具详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "toolSquareId",
                        "name": "toolSquareId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "组织ID",
                        "name": "orgID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ToolSquareDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "mp_common.LLMReq": {
            "type": "object",
            "required": [
                "messages",
                "model"
            ],
            "properties": {
                "do_sample": {
                    "description": "Yuanjing",
                    "type": "boolean"
                },
                "enable_thinking": {
                    "type": "boolean"
                },
                "extra_body": {
                    "description": "扩展参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp_common.ExtraBody"
                        }
                    ]
                },
                "frequency_penalty": {
                    "description": "频率惩罚系数",
                    "type": "number"
                },
                "logit_bias": {
                    "description": "调整指定 token 在模型输出内容中出现的概率",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "logprobs": {
                    "description": "是否返回输出 Token 的对数概率",
                    "type": "boolean"
                },
                "max_completion_tokens": {
                    "description": "控制模型输出的最大长度[0,64k]",
                    "type": "integer"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mp_common.OpenAIReqMsg"
                    }
                },
                "min_p": {
                    "type": "number"
                },
                "model": {
                    "description": "general",
                    "type": "string"
                },
                "n": {
                    "type": "integer"
                },
                "parallel_tool_calls": {
                    "description": "是否开启并行工具调用",
                    "type": "boolean"
                },
                "presence_penalty": {
                    "description": "控制模型生成文本时的内容重复度",
                    "type": "number"
                },
                "repetition_penalty": {
                    "description": "模型生成时连续序列中的重复度",
                    "type": "number"
                },
                "response_format": {
                    "$ref": "#/definitions/mp_common.OpenAIResponseFormat"
                },
                "seed": {
                    "description": "种子",
                    "type": "integer"
                },
                "stop": {
                    "type": "string"
                },
                "stream": {
                    "type": "boolean"
                },
                "stream_options": {
                    "description": "当启用流式输出时，可通过将本参数设置为{\"include_usage\": true}，在输出的最后一行显示所使用的Token数。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp_common.StreamOptions"
                        }
                    ]
                },
                "temperature": {
                    "type": "number"
                },
                "thinking": {
                    "description": "custom",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp_common.Thinking"
                        }
                    ]
                },
                "thinking_budget": {
                    "description": "思考过程的最大长度，只在enable_thinking为true时生效",
                    "type": "integer"
                },
                "tool_choice": {
                    "description": "强制指定工具调用的策略"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mp_common.OpenAITool"
                    }
                },
                "top_k": {
                    "type": "integer"
                },
                "top_logprobs": {
                    "description": "指定在每一步生成时，返回模型最大概率的候选 Token 个数",
                    "type": "integer"
                },
                "top_p": {
                    "type": "number"
                },
                "user": {
                    "type": "string"
                },
                "web_search": {
                    "description": "搜索增强",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp_common.WebSearch"
                        }
                    ]
                }
            }
        },
        "mp_common.LLMResp": {
            "type": "object",
            "required": [
                "choices",
                "model"
            ],
            "properties": {
                "choices": {
                    "description": "生成结果列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mp_common.OpenAIRespChoice"
                    }
                },
                "code": {
                    "type": "integer"
                },
                "created": {
                    "description": "时间戳（秒）",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "img_id": {
                    "description": "视觉模型返回图片id",
                    "type": "string"
                },
                "model": {
                    "description": "使用的模型",
                    "type": "string"
                },
                "object": {
                    "description": "固定为 \"chat.completion\"",
                    "type": "string"
                },
                "service_tier": {
                    "description": "（火山）指定是否使用TPM保障包。生效对象为购买了保障包推理接入点",
                    "type": "string"
                },
                "system_fingerprint": {
                    "type": "string"
                },
                "usage": {
                    "description": "token 使用统计",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp_common.OpenAIRespUsage"
                        }
                    ]
                }
            }
        },
...
        "request.CallbackMetaData": {
            "type": "object",
            "required": [
                "metaId",
                "value"
            ],
            "properties": {
                "key": {
                    "type": "string"
                },
                "metaId": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "request.CallbackUpdateDocStatusReq": {
            "type": "object",
            "required": [
                "id",
                "status"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "metaDataList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.CallbackMetaData"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "request.SearchKnowledgeInfoReq": {
            "type": "object",
            "required": [
                "categoryName",
                "userId"
            ],
            "properties": {
                "categoryName": {
                    "type": "string"
                },
                "orgId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "response.ModelInfo": {
            "type": "object",
            "required": [
                "model",
                "modelType",
                "provider"
            ],
            "properties": {
                "avatar": {
                    "description": "模型图标路径",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.Avatar"
                        }
                    ]
                },
                "config": {},
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "description": "模型显示名称",
                    "type": "string"
                },
                "examples": {
                    "description": "仅用于swagger展示；模型对应供应商中的对应llm、embedding或rerank结构是config实际的参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mp.ProviderModelConfig"
                        }
                    ]
                },
                "isActive": {
                    "description": "启用状态（true: 启用，false: 禁用）",
                    "type": "boolean"
                },
                "model": {
                    "description": "模型名称",
                    "type": "string"
                },
                "modelDesc": {
                    "type": "string"
                },
                "modelId": {
                    "type": "string"
                },
                "modelType": {
                    "type": "string",
                    "enum": [
                        "llm",
                        "embedding",
                        "rerank"
                    ]
                },
                "orgId": {
                    "type": "string"
                },
                "provider": {
                    "description": "模型供应商",
                    "type": "string",
                    "enum": [
                        "OpenAI-API-compatible",
                        "YuanJing"
                    ]
                },
                "publishDate": {
                    "description": "模型发布时间",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mp_common.Tag"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}